MyInfo/MyInfo/textProcesser/sdk.python-master-00b4ee9e94afc191c039701480f10b1a4e964e25/setup.py
import ez_setup
ez_setup.use_setuptools()
from setuptools import setup

setup(
  name='ispras',
  version='1.0.2',
  packages=['ispras'],
  install_requires=[
      'requests',
      'xmltodict'
  ],
  py_modules=['ispras'],
)
