MyInfo/MyInfo/textProcesser/sdk.python-master-00b4ee9e94afc191c039701480f10b1a4e964e25/requirements-test.txt
requests==2.7.0
xmltodict==0.9.2
python-dotenv==0.1.3
